+++ Core/SoarKernel/src/SQLiteCpp/Statement.cpp
@@ -142,7 +142,7 @@ void Statement::bind(const int aIndex, const uint64_t aValue)
 }
 
 // Bind a 64bits int value to a parameter "?", "?NNN", ":VVV", "@VVV" or "$VVV" in the SQL prepared statement
-void Statement::bind(const int aIndex, const long long aValue)
+void Statement::bind(const int aIndex, const int64_t aValue)
 {
     const int ret = sqlite3_bind_int64(mStmtPtr, aIndex, aValue);
     check(ret);
@@ -224,7 +224,7 @@ void Statement::bind(const char* apName, const unsigned aValue)
 }
 
 // Bind a 64bits int value to a parameter "?NNN", ":VVV", "@VVV" or "$VVV" in the SQL prepared statement
-void Statement::bind(const char* apName, const long long aValue)
+void Statement::bind(const char* apName, const int64_t aValue)
 {
     const int index = sqlite3_bind_parameter_index(mStmtPtr, apName);
     const int ret = sqlite3_bind_int64(mStmtPtr, index, aValue);
